+++ entity/src/entity.rs
@@ -1,0 +1,10 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+pub struct EntityId(pub(crate) usize);
+
+pub struct Entities {}
+
+impl Entities {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
