+++ entity/src/system.rs
@@ -1,0 +1,7 @@
+pub struct Systems {}
+
+impl Systems {
+    pub fn new() -> Self {
+        Systems {}
+    }
+}
